application_service/requirements.txt
tensorflow
streamlit
pillow
pandas
matplotlib
numpy
requests
scikit-learn
pyvis